PPC32 bugs when an atomic operation is done on a i8. Change to i32.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@74173 91177308-0d34-0410-b5e6-96231b3b80d8

+++ include/mvm/Threads/Locks.h
@@ -342,7 +342,7 @@ public:
 
   /// locked - Is the spin lock locked?
   ///
-  uint8 locked;
+  uint32 locked;
   
   /// SpinLock - Initialize the lock as not being held.
   ///
